commit 27

+++ client/src/components/ProfileUser.vue
@@ -1,4 +1,5 @@
 <template>
+    <Header />
     <div class="container-profile">
         <div class="block-profile">
             <div class="form-container">
@@ -6,8 +7,15 @@
                 <form @submit.prevent="updateUserData()"
                     style="padding-top: 30px; display: flex; flex-direction: row; gap: 20px; width: auto;">
                     <div class="form-avatar-container">
-                        <img :src="avatarIcon" class="avatar-icon" alt="User Avatar">
-                        <input type="file" @change="handleFileChange" accept="image/*">
+                        <div class="avatar-icon">
+                            <img :src="avatarIcon"  alt="User Avatar">
+
+                                <label class="upload" name="upload" >
+                                    <input type="file" style="display: none;" @change="handleFileChange" accept="image/*">
+                                    <img :src="upload" for="upload" name="upload" alt="">
+                                </label>
+                            
+                        </div>
                     </div>
                     <div class="form-child-container">
                         <div class="profile-group">
@@ -27,48 +35,69 @@
                         </div>
                         <div class="profile-group">
                             <label>Новый пароль:</label>
-                            <input type="password" v-model="updatePassword" placeholder="Введите новый пароль">
-                            <p v-if="errors.password" class="error-message">{{ errors.password }}</p>
+                            <input type="password" v-model="updatePassword" placeholder="******">
+                            <p v-if="errors.updatePassword" class="error-message">{{ errors.updatePassword }}</p>
                         </div>
                         <div class="profile-group">
                             <label>Подтверждение пароля:</label>
-                            <input type="password" v-model="confirmPassword" placeholder="Подтвердите пароль">
+                            <input type="password" v-model="confirmPassword" placeholder="******">
                             <p v-if="errors.confirmPassword" class="error-message">{{ errors.confirmPassword }}</p>
                         </div>
                         <div class="profile-group">
                             <label>Дата рождения:</label>
                             <input type="date" v-model="updateBirthDate">
-                            <p v-if="errors.birthDate" class="error-message">{{ errors.birthDate }}</p>
                         </div>
+                        <p v-if="errors.generalError" class="error-message">{{ errors.generalError }}</p>
                         <button type="submit">Обновить данные</button>
                     </div>
                 </form>
             </div>
         </div>
     </div>
+    <Footer />
 </template>
 
 
 <script setup>
 import { reactive, ref } from 'vue';
+import { useRouter } from "vue-router";
+
+const router = useRouter();
+
+import upload from "../assets/images/upload.svg"
+import Header from './Header.vue';
+import Footer from './Footer.vue';
 
 const props = defineProps({
     showAndHideProfile: Function
 });
 
-const userData = JSON.parse(localStorage.getItem("userData"));
+const userData = JSON.parse(sessionStorage.getItem("userData"));
+const users = JSON.parse(localStorage.getItem("users"));
 
+let avatarIcon = ref(userData.avatarIcon || 'src/assets/images/avatarIcon.png');
 const updateUsername = ref(userData.username);
 const updateEmail = ref(userData.email);
+const updateFullName = ref(userData.fullName || "");
+const updatePassword = ref("");
+const confirmPassword = ref("");
+const updateBirthDate = ref(userData.birthDate || "");
 
 const errors = reactive({
     username: "",
     email: "",
+    updatePassword: "",
+    confirmPassword: "",
+    generalError: ""
 });
 
 const validateData = () => {
     errors.username = "";
     errors.email = "";
+    errors.updatePassword = "";
+    errors.confirmPassword = "";
+    errors.generalError = "";
+
     if (!updateUsername.value.trim()) {
         errors.username = "Это поле обязательно для заполнения";
     }
@@ -79,9 +108,31 @@ const validateData = () => {
     } else if (!emailPattern.test(updateEmail.value)) {
         errors.email = "Некорректная формат почты";
     }
-};
+    
+    if(updatePassword.value || confirmPassword.value){
+        if(updatePassword.value.trim().length < 6){
+            errors.updatePassword = "Минимальная длина пароля 6 символов";
+        }
 
-let avatarIcon = ref(userData.avatarIcon || 'src/assets/images/avatarIcon.png');
+        if(confirmPassword.value.trim().length < 6){
+            errors.confirmPassword = "Минимальная длина пароля 6 символов";
+        }
+
+        if(updatePassword.value !== confirmPassword.value){
+            errors.generalError  = "Пароли не совпадают";
+        }
+    }
+
+    users.map(item => {
+        if (item.email == updateEmail.value && item.id != userData.id) {
+            errors.email = "Эта почта уже занят";
+        }
+        
+        if(item.username == updateUsername.value && item.id != userData.id){
+            errors.username = "Этот ник уже занят";
+        }
+    });
+};
 
 const handleFileChange = (event) => {
     const file = event.target.files[0];
@@ -97,29 +148,48 @@ const handleFileChange = (event) => {
 
 const updateUserData = () => {
     validateData();
-
-    if (errors.email || errors.username) return;
-    const users = JSON.parse(localStorage.getItem("users"));
+    if (errors.generalError || errors.email || errors.username || 
+        errors.updatePassword || errors.confirmPassword) return;
     users.map(item => {
         if (item.id == userData.id) {
             item.username = updateUsername.value;
             item.email = updateEmail.value;
             item.avatarIcon = avatarIcon.value;
+            item.password = updatePassword.value;
+            item.dateOfBirth = updateBirthDate.value;
+            item.fullName = updateFullName.value;
         }
     });
     localStorage.setItem("users", JSON.stringify(users));
+    alert("Данные обновлены");
+    router.push("/");
 };
 </script>
 
 <style scoped>
-.block-profile{
-    background-color: rgba(0, 0, 0, 0.4)
-}
 .form-container {
+    width: 100%;
+    position: relative;
     display: flex;
     flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    z-index: 6;
+    margin-block: 60px;
 }
-
+form{
+    position: relative;
+    width: 420px;
+    display: flex;
+    flex-direction: column;
+    gap: 14px;
+    padding: 30px;
+    border-radius: 20px;
+    background-color: #13101B;
+    color: white;
+    box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
+    font-size: 20px;
+}
 .cross {
     position: absolute;
     top: 16px;
@@ -130,22 +200,51 @@ const updateUserData = () => {
 }
 
 .avatar-icon {
-    width: 200px;
-    height: 200px;
+    height: max-content;
+    position: relative;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin: auto;
+}
+.avatar-icon img {
+    width: 460px;
+    height: 460px;
+    object-fit: cover;
     border-radius: 50%;
+    margin: auto;
 }
 
 .form-child-container {
     display: flex;
     flex-direction: column;
     justify-content: space-around;
+    width: 50%;
 }
 
 .form-avatar-container {
+    padding-top: 30px;
     display: flex;
-    flex-direction: column;
-    justify-content: space-around;
-    align-items: center;
-    gap: 10px;
+    position: relative;
+    flex-direction: row;
+    gap: 20px;
+    width: 1000px;
+    height: 700px;
+}
+
+.form-avatar-container .upload {
+    position: absolute;
+    background: #77BE1D;
+    padding: 10px;
+    border-radius: 50%;
+    bottom: 10px;
+    right: 80px;
+}
+
+.form-avatar-container .upload img {
+    padding:30px;
+    width: 50px;
+    height: 50px;
+    cursor: pointer;
 }
 </style>
